feat(i18n): add initial i18m setup and config
