+++ package.json
@@ -77,5 +77,8 @@
 				}
 			]
 		]
+	},
+	"dependencies": {
+		"i18n": "^0.15.1"
 	}
 }
