+++ packages/forms-web-app/src/locales/cy.json
@@ -1,0 +1,5 @@
+{
+	"examinationTimetable": {
+		"title": "I am a Welsh title"
+	}
+}
