+++ packages/forms-web-app/src/locales/en.json
@@ -1,0 +1,5 @@
+{
+	"examinationTimetable": {
+		"title": "I am a English title"
+	}
+}
